requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
Brotli==1.0.9
distlib==0.3.3
Django==3.2.11 
django-rest-framework==0.1.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
filelock==3.4.0
gunicorn==20.1.0
importlib-metadata==4.8.2
packaging==21.3
platformdirs==2.4.0
pyparsing==3.0.7
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.0
virtualenv==20.10.0
webencodings==0.5.1
whitenoise==6.0.0
zipp==3.6.0
django-cors-headers==3.11.0
tzdata==2021.5
djangorestframework==3.13.1
django-summernote==0.8.20.0
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
dj-database-url==0.5.0
psycopg2-binary==2.9.3